01_04_audio/requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
mcp==0.1.0
aioconsole==0.7.0
